The base class for the scene renderer. Common functionality for renderers moved into it
