+++ core/src/com/vych/game/renderer/scenes/GameSceneRenderer.java
@@ -1,16 +1,10 @@
 package com.vych.game.renderer.scenes;
 
-import com.badlogic.gdx.graphics.g2d.Batch;
-import com.vych.game.managers.gameObjects.GameObjectsManager;
-import com.vych.game.managers.gameObjects.entities.core.GameObject;
-import com.vych.game.managers.resources.ResourcesManager;
 import com.vych.game.managers.resources.entities.core.ResourceType;
+import com.vych.game.renderer.core.BasicSceneRenderer;
 import com.vych.game.renderer.core.SceneAsset;
-import com.vych.game.renderer.core.SceneRenderer;
 
-import java.util.List;
-
-public class GameSceneRenderer implements SceneRenderer {
+public class GameSceneRenderer extends BasicSceneRenderer {
     private final SceneAsset[] assets = {
             new SceneAsset()
                     .setResourceName("dropImage")
@@ -40,28 +34,11 @@ public class GameSceneRenderer implements SceneRenderer {
 
     @Override
     public void loadAssets() {
-        ResourcesManager resourcesManager = ResourcesManager.getInstance();
-        for (SceneAsset asset : this.assets) {
-            resourcesManager.loadResource(asset);
-        }
+        loadAssets(this.assets);
     }
 
     @Override
     public void unloadAssets() {
-        ResourcesManager resourcesManager = ResourcesManager.getInstance();
-        for (SceneAsset asset : this.assets) {
-            resourcesManager.unloadResource(asset);
-        }
-    }
-
-    @Override
-    public void renderScene(Batch batch) {
-        List<GameObject> objectsToRender = GameObjectsManager.getInstance().getObjectsByRenderer(this);
-        for (GameObject obj : objectsToRender) {
-            obj.instanceRender(batch);
-        }
-        for (GameObject obj : objectsToRender) {
-            obj.instanceRenderGUI(batch);
-        }
+        unloadAssets(this.assets);
     }
 }
